move metadata helpers to state.ts and fix forward bug
